refactor(api): use inspect.signature in command publish
inspect.getfullargspec is a legacy API meant to mimic the behavior of python 2.
It is missing some key features, like follow_wrapped, from inspect.signature.

Use inspect.signature here so we can use wrapped functions in publish args. This
also lets us have the library do more of the argument binding, with the downside
that we need to be a bit more careful about whether the method we're talking
about is a free function or a bound method.

We also need to fix up some misuses and bugs that this change illuminated.